applied [comment](https://github.com/pendulum-chain/spacewalk/pull/98#pullrequestreview-1149717484)
